feat: remove spacing in full screen compact mode
Spacing needs to be 1 for compact mode when the context pane is visible.
This is due to a quirk in how the list bubble renders. The list renders
with 1 extra line in this case, which pushes the whole view up, causing
UI inconsistencies. Handling this case by investigating the state of the
model has been frustratingly futile, so spacing stays at 1 for this view
for now.